+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Updated the Product list so that recipe families/variants (e.g., "Board",
   "Hardwood Board", and "Softwood Board") are grouped together.
+- Added the ability to filter the Products lists by Blueprints.
 
 ## 0.1.0 (2026-05-02)